64bits fixes.
        * capplets/font/main.c: (sample_expose):
        * gnome-settings-daemon/gnome-settings-gtk1theme.c:
        (apply_settings):
        * vfs-methods/themus/theme-method.c:
        64bits fixes.

+++ gnome-settings-daemon/gnome-settings-gtk1theme.c
@@ -151,7 +151,7 @@ apply_settings (void)
 	gchar *theme_filename;
 	gchar *rc_filename;
 	gchar *current_contents;
-	gint current_length;
+	gsize current_length;
 	gchar *new_contents;
 	GError *err = NULL;
